Feature band fallback with workload update

+++ src/Cli/dotnet/commands/dotnet-workload/install/WorkloadManifestUpdater.cs
@@ -81,7 +81,8 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
 
         public async Task UpdateAdvertisingManifestsAsync(bool includePreviews, DirectoryPath? offlineCache = null)
         {
-            var manifests = GetInstalledManifestIds();
+            // this updates all the manifests 
+            var manifests = _workloadResolver.GetInstalledManifests();
             await Task.WhenAll(manifests.Select(manifest => UpdateAdvertisingManifestAsync(manifest, includePreviews, offlineCache)))
                 .ConfigureAwait(false);
             WriteUpdatableWorkloadsFile();
@@ -184,7 +185,9 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
                 }
 
                 if (advertisingManifestVersionAndWorkloads != null &&
-                    advertisingManifestVersionAndWorkloads.Value.ManifestVersion.CompareTo(currentManifestVersion.Item1) > 0)
+                    ((advertisingManifestVersionAndWorkloads.Value.ManifestVersion.CompareTo(currentManifestVersion.manifestVersion) > 0 
+                        && advertisingManifestVersionAndWorkloads.Value.ManifestFeatureBand.Equals(currentManifestVersion.sdkFeatureBand)) ||
+                    advertisingManifestVersionAndWorkloads.Value.ManifestFeatureBand.CompareTo(currentManifestVersion.sdkFeatureBand) > 0)) 
                 {
                     manifestUpdates.Add((new ManifestVersionUpdate(manifestId, currentManifestVersion.manifestVersion, currentManifestVersion.sdkFeatureBand.ToString(),
                         advertisingManifestVersionAndWorkloads.Value.ManifestVersion, advertisingManifestVersionAndWorkloads.Value.ManifestFeatureBand.ToString()),
@@ -292,40 +295,33 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
             return _workloadResolver.GetInstalledManifests().Select(manifest => new ManifestId(manifest.Id));
         }
 
-        private async Task UpdateAdvertisingManifestAsync(ManifestId manifestId, bool includePreviews, DirectoryPath? offlineCache = null)
+        private async Task UpdateAdvertisingManifestAsync(WorkloadManifestInfo manifest, bool includePreviews, DirectoryPath? offlineCache = null)
         {
             string packagePath = null;
             string extractionPath = null;
+            var manifestId = new ManifestId(manifest.Id);
+            string currentFeatureBand = _sdkFeatureBand.ToString();
 
             try
             {
                 var adManifestPath = GetAdvertisingManifestPath(_sdkFeatureBand, manifestId);
-                if (offlineCache == null || !offlineCache.HasValue)
+                
+                bool success;
+                (success, packagePath) = await GetManifestPackageUpdate(_sdkFeatureBand, manifestId, includePreviews, offlineCache);
+                if (!success)
                 {
-                    try
+                    if (!(manifest.ManifestFeatureBand).Equals(_sdkFeatureBand))
                     {
-                        packagePath = await _nugetPackageDownloader.DownloadPackageAsync(
-                            GetManifestPackageId(_sdkFeatureBand, manifestId),
-                            packageSourceLocation: _packageSourceLocation,
-                            includePreview: includePreviews);
-                    }
-                    catch (NuGetPackageNotFoundException)
-                    {
-                        _reporter.WriteLine(string.Format(LocalizableStrings.AdManifestPackageDoesNotExist, manifestId));
+                        (success, packagePath) = await GetManifestPackageUpdate(new SdkFeatureBand(manifest.ManifestFeatureBand), manifestId, includePreviews, offlineCache);
+                        currentFeatureBand = manifest.ManifestFeatureBand.ToString();
                     }
                 }
-                else
+                if (!success)
                 {
-                    packagePath = Directory.GetFiles(offlineCache.Value.Value)
-                        .Where(path => path.EndsWith(".nupkg"))
-                        .Where(path => Path.GetFileName(path).StartsWith(GetManifestPackageId(_sdkFeatureBand, manifestId).ToString()))
-                        .Max();
-                    if (!File.Exists(packagePath))
-                    {
-                        throw new Exception(string.Format(LocalizableStrings.CacheMissingPackage, GetManifestPackageId(_sdkFeatureBand, manifestId), "*", offlineCache));
-                    }
+                    _reporter.WriteLine(string.Format(LocalizableStrings.AdManifestPackageDoesNotExist, manifestId));
+                    return;
                 }
-
+                
                 extractionPath = Path.Combine(_tempDirPath, "dotnet-sdk-advertising-temp", $"{manifestId}-extracted");
                 Directory.CreateDirectory(extractionPath);
                 var resultingFiles = await _nugetPackageDownloader.ExtractPackageAsync(packagePath, new DirectoryPath(extractionPath));
@@ -336,6 +332,9 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
                 }
                 Directory.CreateDirectory(Path.GetDirectoryName(adManifestPath));
                 FileAccessRetrier.RetryOnMoveAccessFailure(() => DirectoryPath.MoveDirectory(Path.Combine(extractionPath, "data"), adManifestPath));
+
+                // add file that contains the advertisted manifest feature band so GetAdvertisingManifestVersionAndWorkloads will use correct feature band, regardless of if rollback occurred or not
+                File.WriteAllText(Path.Combine(adManifestPath, "AdvertisedManifestFeatureBand.txt"), currentFeatureBand);
 
                 if (_displayManifestUpdates)
                 {
@@ -358,15 +357,18 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
                 {
                     File.Delete(packagePath);
                 }
-
-                var versionDir = Path.GetDirectoryName(packagePath);
-                if (Directory.Exists(versionDir) && !Directory.GetFileSystemEntries(versionDir).Any())
+                if (!string.IsNullOrEmpty(packagePath) && (offlineCache == null || !offlineCache.HasValue))
                 {
-                    Directory.Delete(versionDir);
-                    var idDir = Path.GetDirectoryName(versionDir);
-                    if (Directory.Exists(idDir) && !Directory.GetFileSystemEntries(idDir).Any())
+                    var versionDir = Path.GetDirectoryName(packagePath);
+
+                    if (Directory.Exists(versionDir) && !Directory.GetFileSystemEntries(versionDir).Any())
                     {
-                        Directory.Delete(idDir);
+                        Directory.Delete(versionDir);
+                        var idDir = Path.GetDirectoryName(versionDir);
+                        if (Directory.Exists(idDir) && !Directory.GetFileSystemEntries(idDir).Any())
+                        {
+                            Directory.Delete(idDir);
+                        }
                     }
                 }
             }
@@ -385,10 +387,18 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
             using (FileStream fsSource = new FileStream(manifestPath, FileMode.Open, FileAccess.Read))
             {
                 var manifest = WorkloadManifestReader.ReadWorkloadManifest(manifestId.ToString(), fsSource, manifestPath);
+                // we need to know the feature band of the advertised manifest (read it from the AdvertisedManifestFeatureBand.txt file)
+                // if we don't find the file then use the current feature band
+                var adManifestFeatureBandPath = Path.Combine(GetAdvertisingManifestPath(_sdkFeatureBand, manifestId), "AdvertisedManifestFeatureBand.txt");
 
-                //  TODO: figure out how to differentiate between the feature band an advertising manifest is branded as and the feature band of the SDK
-                //  it's advertised to
-                return (new ManifestVersion(manifest.Version), _sdkFeatureBand, manifest.Workloads.Values.OfType<WorkloadDefinition>().ToDictionary(w => w.Id));
+                SdkFeatureBand adManifestFeatureBand = _sdkFeatureBand;
+                if (File.Exists(adManifestFeatureBandPath))
+                {
+                    adManifestFeatureBand = new SdkFeatureBand(File.ReadAllText(adManifestFeatureBandPath));
+                }
+                              
+
+                return (new ManifestVersion(manifest.Version), adManifestFeatureBand, manifest.Workloads.Values.OfType<WorkloadDefinition>().ToDictionary(w => w.Id));
             }
         }
 
@@ -464,7 +474,7 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
                 else
                 {
                     throw new ArgumentException(string.Format(LocalizableStrings.RollbackDefinitionFileDoesNotExist, rollbackDefinitionFilePath));
-                }           
+                }
             }
             return JsonSerializer.Deserialize<IDictionary<string, string>>(fileContent)
                 .Select(manifest =>
@@ -500,6 +510,51 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
         private string GetAdvertisingWorkloadsFilePath(SdkFeatureBand featureBand) => GetAdvertisingWorkloadsFilePath(_userProfileDir, featureBand);
 
         private static string GetAdvertisingWorkloadsFilePath(string userProfileDir, SdkFeatureBand featureBand) => Path.Combine(userProfileDir, $".workloadAdvertisingUpdates{featureBand}");
+
+        private async Task<String> GetOnlinePackagePath(SdkFeatureBand sdkFeatureBand, ManifestId manifestId, bool includePreviews) 
+            {
+                string packagePath = await _nugetPackageDownloader.DownloadPackageAsync(
+                    GetManifestPackageId(sdkFeatureBand, manifestId),
+                    packageSourceLocation: _packageSourceLocation,
+                    includePreview: includePreviews);
+                
+                return packagePath;
+            }
+
+        private string GetOfflinePackagePath(SdkFeatureBand sdkFeatureBand, ManifestId manifestId, DirectoryPath? offlineCache = null)
+        {
+            string packagePath = Directory.GetFiles(offlineCache.Value.Value)
+                .Where(path => path.EndsWith(".nupkg"))
+                .Where(path =>
+                {
+                    var manifestPackageId = GetManifestPackageId(sdkFeatureBand, manifestId).ToString();
+                    return Path.GetFileName(path).StartsWith(manifestPackageId, StringComparison.OrdinalIgnoreCase);
+                })
+                .Max();
+
+            return packagePath;
+        }
+
+        private async Task<(bool, string)> GetManifestPackageUpdate(SdkFeatureBand sdkFeatureBand, ManifestId manifestId, bool includePreviews, DirectoryPath? offlineCache = null)
+        {
+            if (offlineCache == null || !offlineCache.HasValue)
+            {
+                try 
+                {
+                    string packagePath = await GetOnlinePackagePath(sdkFeatureBand, manifestId, includePreviews);
+                    return (true, packagePath);
+                }
+                catch (NuGetPackageNotFoundException)
+                {
+                    return (false, null);
+                }
+            }
+            else
+            {
+                string packagePath = GetOfflinePackagePath(sdkFeatureBand, manifestId, offlineCache);
+                return (packagePath != null, packagePath);
+            }
+        }
 
         private string GetAdvertisingManifestPath(SdkFeatureBand featureBand, ManifestId manifestId) =>
             Path.Combine(_userProfileDir, "sdk-advertising", featureBand.ToString(), manifestId.ToString());
